--- login.py
@@ -1,76 +1,0 @@
-from fastapi import FastAPI, Depends, HTTPException, status
-from sqlalchemy.orm import Session
-from passlib.context import CryptContext
-from pydantic import BaseModel
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from models import Base, SessionLocal, User
-from fastapi.middleware.cors import CORSMiddleware
-
-#pip install passlib
-#pip install uvicorn
-
-app = FastAPI()
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["http://localhost:3000"],  # 모든 접근 허용 -> 개발단계에서만 사용해야됨
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-# 비밀번호 해싱
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-class UserCreate(BaseModel):
-    email: str
-    name: str
-    password: str
-
-class UserLogin(BaseModel):
-    email: str
-    password: str
-
-# 비밀번호 해싱
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
-
-
-@app.post("/signup", status_code=201)
-def signup(user: UserCreate, db: Session = Depends(get_db)):
- 
-    db_user = db.query(User).filter(User.email == user.email).first()
-    if db_user:
-        raise HTTPException(status_code=400, detail="중복된 Email이 있습니다.")
-        #email 중복 메시지
-
-
-    hashed_password = get_password_hash(user.password)
-    new_user = User(email=user.email, name=user.name, password=hashed_password)
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
-
-    return {"message": "회원가입이 정상적으로 실행되었습니다."}
-
-
-@app.post("/login")
-def login(user: UserLogin, db: Session = Depends(get_db)):
-
-    db_user = db.query(User).filter(User.email == user.email).first()
-    if not db_user or not verify_password(user.password, db_user.password):
-        raise HTTPException(status_code=401, detail="로그인 실패")
-
-    return {"message": "로그인 성공"}
