+++ main.py
@@ -1,0 +1,114 @@
+from fastapi import FastAPI, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+from passlib.context import CryptContext
+from pydantic import BaseModel
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+from models import Base, SessionLocal, User
+from fastapi.middleware.cors import CORSMiddleware
+import jwt
+from datetime import datetime, timedelta
+from fastapi.security import OAuth2PasswordBearer
+
+#pip install pyjwt -> jwt설치
+# pip install passlib
+# pip install uvicorn
+# 실행 방법: uvicorn main:app --reload
+
+app = FastAPI()
+
+# CORS 설정
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["http://localhost:3000"],  # 개발 단계에서만 사용
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+# 비밀번호 해싱 설정
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+# JWT 설정
+SECRET_KEY = "shinjunghwan_goat" #jwt 보안 키 
+ALGORITHM = "HS256" #none으로 설정하면 보안이슈 발생함으로 지양
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+# JWT 토큰 생성 함수
+def create_access_token(data: dict):
+    to_encode = data.copy()
+    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+# JWT 토큰 검증 함수
+def verify_token(token: str):
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        return payload
+    except jwt.ExpiredSignatureError:
+        raise HTTPException(status_code=401, detail="토큰 만료")
+    except jwt.InvalidTokenError:
+        raise HTTPException(status_code=401, detail="유효하지 않은 토큰")
+
+# 데이터베이스 세션 생성
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+# 사용자 모델 정의
+class UserCreate(BaseModel):
+    email: str
+    name: str
+    password: str
+
+class UserLogin(BaseModel):
+    email: str
+    password: str
+
+# 비밀번호 해싱 및 검증 함수
+def get_password_hash(password):
+    return pwd_context.hash(password)
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
+
+# OAuth2 설정
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
+
+# 사용자 검증 함수
+def get_current_user(token: str = Depends(oauth2_scheme)):
+    payload = verify_token(token)
+    email = payload.get("sub")
+    if email is None:
+        raise HTTPException(status_code=401, detail="유효하지 않은 사용자입니다.")
+    return email
+
+# 회원가입 엔드포인트
+@app.post("/signup", status_code=201)
+def signup(user: UserCreate, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.email == user.email).first()
+    if db_user:
+        raise HTTPException(status_code=400, detail="중복된 Email이 있습니다.")
+    
+    hashed_password = get_password_hash(user.password)
+    new_user = User(email=user.email, name=user.name, password=hashed_password)
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+
+    return {"message": "회원가입이 정상적으로 실행되었습니다."}
+
+# 로그인 엔드포인트
+@app.post("/login")
+def login(user: UserLogin, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.email == user.email).first()
+    if not db_user or not verify_password(user.password, db_user.password):
+        raise HTTPException(status_code=401, detail="로그인 실패")
+    
+    access_token = create_access_token(data={"sub": db_user.email})
+    return {"access_token": access_token, "token_type": "bearer"}
